refactor: extract shared mock helpers to reduce test duplication
Created tests/unit/import/open5e.mockHelpers.ts with:
- createMockFetch, createMockFetch429 for API mocking
- createPaginatedResponse for paginated data
- SAMPLE_CREATURE, SAMPLE_SPELL constants
- createTestCreature, createTestSpell factory functions
- createMockClient for IOpen5EClient mocking

Refactored test files to use shared helpers:
- open5eAdapter.test.ts - uses shared mocks and sample data
- dedupeEngine.test.ts - uses shared createMockClient
- testFactories.ts - now re-exports from open5e.mockHelpers

This reduces code duplication across import test files and should
improve Codacy duplication metrics.
