+++ tests/unit/import/open5eAdapter.test.ts
@@ -1,46 +1,15 @@
-import { Open5EClient, IOpen5EClient, PaginatedResponse, Open5ECreature, Open5ESpell } from "@/lib/import/open5eAdapter";
+import { Open5EClient, PaginatedResponse, Open5ECreature, Open5ESpell } from "@/lib/import/open5eAdapter";
+import {
+  createMockFetch,
+  createPaginatedResponse,
+  SAMPLE_CREATURE,
+  SAMPLE_SPELL,
+} from "./open5e.mockHelpers";
 
 describe("Open5EClient", () => {
-  function createMockFetch(response: unknown) {
-    return jest.fn().mockResolvedValue({
-      ok: true,
-      json: jest.fn().mockResolvedValue(response),
-      status: 200,
-      headers: { get: jest.fn().mockReturnValue(null) },
-      clone: jest.fn().mockReturnThis(),
-    });
-  }
-
   describe("fetchMonsters", () => {
     it("returns paginated creatures from API", async () => {
-      const mockData: PaginatedResponse<Open5ECreature> = {
-        results: [
-          {
-            key: "goblin",
-            name: "Goblin",
-            size: { Name: "Small", key: "small" },
-            type: { Name: "Humanoid", key: "humanoid" },
-            ability_scores: {
-              strength: 8,
-              dexterity: 14,
-              constitution: 12,
-              intelligence: 10,
-              wisdom: 8,
-              charisma: 8,
-            },
-            hit_points: 7,
-            armor_class: 15,
-            challenge_rating: 0.25,
-            actions: [],
-            speed: { walk: 30 },
-            alignment: "neutral",
-          },
-        ],
-        count: 1,
-        next: null,
-        previous: null,
-      };
-
+      const mockData = createPaginatedResponse<Open5ECreature>([SAMPLE_CREATURE]);
       const mockFetch = createMockFetch(mockData);
       const client = new Open5EClient(mockFetch as unknown as typeof fetch);
 
@@ -53,13 +22,7 @@ describe("Open5EClient", () => {
     });
 
     it("passes page parameter correctly", async () => {
-      const mockData: PaginatedResponse<Open5ECreature> = {
-        results: [],
-        count: 0,
-        next: null,
-        previous: null,
-      };
-
+      const mockData = createPaginatedResponse<Open5ECreature>([]);
       const mockFetch = createMockFetch(mockData);
       const client = new Open5EClient(mockFetch as unknown as typeof fetch);
 
@@ -73,29 +36,7 @@ describe("Open5EClient", () => {
 
   describe("fetchSpells", () => {
     it("returns paginated spells from API", async () => {
-      const mockData: PaginatedResponse<Open5ESpell> = {
-        results: [
-          {
-            key: "fireball",
-            name: "Fireball",
-            level: 3,
-            school: { Name: "Evocation", key: "evocation" },
-            concentration: false,
-            casting_time: "1 action",
-            range: 0,
-            range_text: "Self",
-            duration: "Instantaneous",
-            verbal: true,
-            somatic: true,
-            material: true,
-            desc: "A ball of fire",
-          },
-        ],
-        count: 1,
-        next: null,
-        previous: null,
-      };
-
+      const mockData = createPaginatedResponse<Open5ESpell>([SAMPLE_SPELL]);
       const mockFetch = createMockFetch(mockData);
       const client = new Open5EClient(mockFetch as unknown as typeof fetch);
 
@@ -110,57 +51,25 @@ describe("Open5EClient", () => {
 
   describe("getAllMonsters", () => {
     it("yields all monsters from all pages", async () => {
-      const page1: PaginatedResponse<Open5ECreature> = {
-        results: [{
-          key: "goblin",
-          name: "Goblin",
-          size: { Name: "Small", key: "small" },
-          type: { Name: "Humanoid", key: "humanoid" },
-          ability_scores: {
-            strength: 8,
-            dexterity: 14,
-            constitution: 12,
-            intelligence: 10,
-            wisdom: 8,
-            charisma: 8,
-          },
-          hit_points: 7,
-          armor_class: 15,
-          challenge_rating: 0.25,
-          actions: [],
-          speed: { walk: 30 },
-          alignment: "neutral",
-        }],
-        count: 2,
-        next: "https://api.open5e.com/v2/creatures/?page=2",
-        previous: null,
+      const goblin: Open5ECreature = { ...SAMPLE_CREATURE, key: "goblin", name: "Goblin" };
+      const orc: Open5ECreature = {
+        ...SAMPLE_CREATURE,
+        key: "orc",
+        name: "Orc",
+        ability_scores: {
+          strength: 16,
+          dexterity: 12,
+          constitution: 16,
+          intelligence: 7,
+          wisdom: 11,
+          charisma: 10,
+        },
       };
 
-      const page2: PaginatedResponse<Open5ECreature> = {
-        results: [{
-          key: "orc",
-          name: "Orc",
-          size: { Name: "Medium", key: "medium" },
-          type: { Name: "Humanoid", key: "humanoid" },
-          ability_scores: {
-            strength: 16,
-            dexterity: 12,
-            constitution: 16,
-            intelligence: 7,
-            wisdom: 11,
-            charisma: 10,
-          },
-          hit_points: 15,
-          armor_class: 13,
-          challenge_rating: 0.5,
-          actions: [],
-          speed: { walk: 30 },
-          alignment: "chaotic evil",
-        }],
-        count: 2,
-        next: null,
-        previous: null,
-      };
+      const page1 = createPaginatedResponse<Open5ECreature>([goblin], {
+        next: "https://api.open5e.com/v2/creatures/?page=2",
+      });
+      const page2 = createPaginatedResponse<Open5ECreature>([orc]);
 
       let callCount = 0;
       const mockFetch = jest.fn().mockImplementation(() => {
@@ -188,33 +97,8 @@ describe("Open5EClient", () => {
     });
 
     it("stops when no more pages", async () => {
-      const singlePage: PaginatedResponse<Open5ECreature> = {
-        results: [{
-          key: "goblin",
-          name: "Goblin",
-          size: { Name: "Small", key: "small" },
-          type: { Name: "Humanoid", key: "humanoid" },
-          ability_scores: {
-            strength: 8,
-            dexterity: 14,
-            constitution: 12,
-            intelligence: 10,
-            wisdom: 8,
-            charisma: 8,
-          },
-          hit_points: 7,
-          armor_class: 15,
-          challenge_rating: 0.25,
-          actions: [],
-          speed: { walk: 30 },
-          alignment: "neutral",
-        }],
-        count: 1,
-        next: null,
-        previous: null,
-      };
-
-      const mockFetch = createMockFetch(singlePage);
+      const mockData = createPaginatedResponse<Open5ECreature>([SAMPLE_CREATURE]);
+      const mockFetch = createMockFetch(mockData);
       const client = new Open5EClient(mockFetch as unknown as typeof fetch);
       const monsters: Open5ECreature[] = [];
 
@@ -229,45 +113,18 @@ describe("Open5EClient", () => {
 
   describe("getAllSpells", () => {
     it("yields all spells from all pages", async () => {
-      const page1: PaginatedResponse<Open5ESpell> = {
-        results: [{
-          key: "fireball",
-          name: "Fireball",
-          level: 3,
-          school: { Name: "Evocation", key: "evocation" },
-          concentration: false,
-          casting_time: "1 action",
-          range: 0,
-          range_text: "Self",
-          duration: "Instantaneous",
-          verbal: true,
-          somatic: true,
-          desc: "Fire",
-        }],
-        count: 2,
-        next: "https://api.open5e.com/v2/spells/?page=2",
-        previous: null,
+      const fireball: Open5ESpell = { ...SAMPLE_SPELL, key: "fireball", name: "Fireball" };
+      const magicMissile: Open5ESpell = {
+        ...SAMPLE_SPELL,
+        key: "magic-missile",
+        name: "Magic Missile",
+        level: 1,
       };
 
-      const page2: PaginatedResponse<Open5ESpell> = {
-        results: [{
-          key: "magic-missile",
-          name: "Magic Missile",
-          level: 1,
-          school: { Name: "Evocation", key: "evocation" },
-          concentration: false,
-          casting_time: "1 action",
-          range: 0,
-          range_text: "Self",
-          duration: "Instantaneous",
-          verbal: true,
-          somatic: true,
-          desc: " missiles",
-        }],
-        count: 2,
-        next: null,
-        previous: null,
-      };
+      const page1 = createPaginatedResponse<Open5ESpell>([fireball], {
+        next: "https://api.open5e.com/v2/spells/?page=2",
+      });
+      const page2 = createPaginatedResponse<Open5ESpell>([magicMissile]);
 
       let callCount = 0;
       const mockFetch = jest.fn().mockImplementation(() => {
